fix(general): mejoras generales de funcionalidades, manejo de pagos multiples y mejoras de diseño
