+++ Backend/src/interfaces/sales.interfaces.ts
@@ -2,7 +2,7 @@ export interface SaleCreate {
     customer: string;
     total: number;
     paymentmethod: string;
-    paymentbreakdown?: { method: string; amount: number }[]; // Desglose de múltiples métodos de pago
+    paymentbreakdown?: { method: string; amount: number }[];
     products: any;
     time: Date;
     ganancias: string;
