+++ Frontend/src/app/core/models/index.ts
@@ -6,3 +6,4 @@ export type { Product } from './product.interface';
 export type { Category, PaymentMethod } from './settings.interface';
 export type { DashboardData, productsTop } from './dashboard.interface';
 export type { User, createUser } from './users.interface';
+export type { Sale } from './sales.interface';
